move /app/lib to /app/utils
